Hard coded in the positions of the first and last nodes into the pos dict
Ok I’ve fixed the “node 0 has no position” error (basically it wasn’t creating a value for the coords of node 0 in the pos dict if node 0 wasn’t connecter to anything), but ofc in these graphs there is no path from start to finish so we get a NoPath error

I've also noticed (for the random seed used in this commit in Graph) that we can get a "node 1 has no position" error if the radius is too small. This seems to suggest something wrong in the logic in the for loops in the add_edges method which populate pos (but only with nodes that are connected, and we can see in the plot that node 1 is very far away).
Perhaps this can be solved by assigning the positions outside of the add_edges method